Cast the gx color index to an int inside the switch argument to avoid problems with the HP-UX 10 compiler which does not accept 64 bit arguments to switch(). Bug 687804.
git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@5524 a1074d23-0009-0410-80fe-cf8c14f379e6

+++ gs/src/gdevstc.c
@@ -2411,7 +2411,8 @@ stc_map_color_cmyk10(gx_device *pdev, gx_color_index color,
 
    c    =   stc_expand(sd,3,(color>>2)&0x3ff);
 
-   switch(color & 3) {
+   /* cast the 64 bit switch argument to work around broken HPUX 10 cc */
+   switch((int)(color & 3)) {
      case 0:
         m = stc_expand(sd,1,(color>>22) & 0x3ff);
         y = stc_expand(sd,2,(color>>12) & 0x3ff);
